+++ lib/graphlient/adapters/http/faraday_adapter.rb
@@ -1,4 +1,5 @@
 require 'faraday'
+require 'json'
 
 module Graphlient
   module Adapters
@@ -13,7 +14,8 @@ module Graphlient
               variables: variables
             }.to_json
           end
-          response.body
+
+          parse_body(response.body)
         rescue Faraday::ConnectionFailed => e
           raise Graphlient::Errors::ConnectionFailedError, e
         rescue Faraday::TimeoutError => e
@@ -37,6 +39,38 @@ module Graphlient
             else
               c.adapter Faraday::Adapter::NetHttp
             end
+          end
+        end
+
+        private
+
+        # Faraday 2.x's JSON response middleware will only parse a JSON
+        # response body into a Hash (or Array) object if the response headers
+        # match a specific content type regex. See Faraday's response JSON
+        # middleware definition for specifics on what the datatype of the
+        # response body will be. This method will handle the situation where
+        # the response header is not set appropriately, but contains JSON
+        # anyways. If the body cannot be parsed as JSON, an exception will be
+        # raised.
+        def parse_body(body)
+          case body
+          when Hash, Array
+            body
+          when String
+            begin
+              JSON.parse(body)
+            rescue JSON::ParserError
+              raise Graphlient::Errors::ServerError, 'Failed to parse response body as JSON'
+            end
+          else
+            inner_exception = StandardError.new <<~ERR.strip.tr("\n", ' ')
+              Unexpected response body type '#{body.class}'. Graphlient doesn't
+              know how to handle a response body of this type, but Faraday is
+              returning it. Please open an issue, particularly if the response
+              body does actually contain valid JSON.
+            ERR
+
+            raise Graphlient::Errors::ClientError, inner_exception
           end
         end
       end